requirements.txt
nose
requests
textblob
python-dateutil
Flask
flask-cors
beautifulsoup4
nltk
